add jwt(json web token) to the system for security purpose.

+++ resources/item.py
@@ -1,13 +1,14 @@
-from flask_smorest  import abort, Blueprint
 from flask.views import MethodView
+from flask_jwt_extended import jwt_required, get_jwt
+from flask_smorest import abort, Blueprint
 from sqlalchemy.exc import SQLAlchemyError
 
 from db import db
-
 from models import ItemModel
 from schemas import ItemSchema, ItemUpdateSchema
 
 blp = Blueprint("Items", __name__, description="Operations on items")
+
 
 @blp.route("/item/<int:item_id>")
 class Items(MethodView):
@@ -16,7 +17,11 @@ class Items(MethodView):
         item = ItemModel.query.get_or_404(item_id)
         return item
 
+    @jwt_required(fresh=True)
     def delete(self, item_id):
+        jwt = get_jwt()
+        if not jwt.get("is_admin"):
+            abort(401, message="Admin privilege required.")
         item = ItemModel.query.get_or_404(item_id)
         db.session.delete(item)
         db.session.commit()
@@ -30,7 +35,7 @@ class Items(MethodView):
             item.price = item_data["price"]
             item.name = item_data['name']
         else:
-            item = ItemModel(id = item_id, **item_data)
+            item = ItemModel(id=item_id, **item_data)
 
         db.session.add(item)
         db.session.commit()
@@ -40,10 +45,12 @@ class Items(MethodView):
 
 @blp.route("/item")
 class ItemsList(MethodView):
+    @jwt_required()
     @blp.response(200, ItemSchema(many=True))
     def get(self):
         return ItemModel.query.all()
 
+    @jwt_required()
     @blp.arguments(ItemSchema)
     @blp.response(201, ItemSchema)
     def post(self, item_data):
@@ -53,5 +60,5 @@ class ItemsList(MethodView):
             db.session.add(item)
             db.session.commit()
         except SQLAlchemyError:
-            abort(500, message="An error occured while inserting the item.")
+            abort(500, message="An error occurred while inserting the item.")
         return item
